move website pages into separate folder
