+++ __tests__/home.tsx
@@ -1,4 +1,4 @@
-import Page from "@/app/page";
+import Page from "@/app/(website)/page";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
