+++ src/app/(website)/journey/page.tsx
